KYC/Image matching from db/image-database-comparison-main/image-database-comparison-main/requirement.txt
numpy 
pillow
matplotlib 
tensorflow 
scikit-learn 
annoy 
tqdm
joblib
streamlit
